% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pal}
\alias{pal}
\title{Color ramp based on \code{mblabs} theme colors}
\usage{
pal(x, ...)
}
\arguments{
\item{x}{number of colors to interpolate}

\item{...}{arguments to pass to \code{colorRamp}.}
}
\value{
A function to interpolate colors
}
\description{
Default qualitative color ramp derived from \code{mblabs} Bootstrap theme.
}
\examples{

x <- rchisq(100, df=4)
hist(x, freq=FALSE, ylim=c(0, 0.2), col=pal(20), border=12)
hist(x, freq=FALSE, ylim=c(0, 0.2), col=pal(8), border=12)

}
