% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pal12}
\alias{pal12}
\title{\code{mblabs} 12-color palette}
\format{
A named character vector of hex color codes
}
\usage{
pal12
}
\description{
Default named qualitative color palette with 12 hues derived from \code{mblabs}
Bootstrap theme.
}
\keyword{datasets}
