% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_labs_on}
\alias{theme_labs_on}
\title{Apply graphic color palettes}
\usage{
theme_labs_on(
  bg = "transparent",
  fg = "#333333",
  accent = pal12[["teal"]],
  font = "Roboto Condensed",
  sequential = sequential_gradient(fg_weight = 0.5, bg_weight = 0.5, fg_low = FALSE),
  qualitative = alpha(pal(12), 0.9),
  ...
)
}
\arguments{
\item{bg}{a background color.}

\item{fg}{a foreground color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{font}{a \code{\link[thematic:font_spec]{font_spec()}} object. If missing, font defaults are not altered.}

\item{sequential}{a color palette for graphical markers that encode
numeric values. Can be a vector of color codes or a
\code{\link[thematic:sequential_gradient]{sequential_gradient()}} object.}

\item{qualitative}{a color palette for graphical markers that encode
qualitative values (won't be used in ggplot2 when the number of data
levels exceeds the max allowed colors). Defaults to \code{\link[thematic:okabe_ito]{okabe_ito()}}.}

\item{...}{
  Arguments passed on to \code{\link[thematic:thematic_on]{thematic::thematic_on}}
  \describe{
    \item{\code{inherit}}{should non-specified values inherit from the previous theme?}
  }}
}
\value{
A global graphic theme
}
\description{
Sets \code{mblabs} default values to \code{thematic::thematic_on}, which modify
\code{lattice} and \code{ggplot} color palettes and fonts.
}
\examples{
theme_labs_on()
lattice::show.settings()

}
