% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yday_to_date}
\alias{yday_to_date}
\title{Format yday dates}
\usage{
yday_to_date(x, format = "\%m-\%d")
}
\arguments{
\item{x}{A numeric vector to format}

\item{format}{\code{\link[base]{character}} string.  If not specified, it will try
    \code{tryFormats} one by one on the first non-\code{NA} element, and
    give an error if none works.  Otherwise, the processing is via
    \code{\link[base]{strptime}()} whose help page describes available
    conversion specifications.}
}
\value{
formatted character vector
}
\description{
Format yday dates
}
